image-classifier/requirements.txt
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
Pillow>=10.0.0
PyYAML>=6.0.0
tqdm>=4.66.0
streamlit>=1.33.0
